Some typos fixed with the PlayerJoinedEvent and added gitignore file
