+++ .gitignore
@@ -1,0 +1,5 @@
+**/.vs
+**/obj
+**/bin
+*.suo
+*.user
